test: co-locate app module tests
Move the monolithic TUI app test coverage into the implementation
modules that own each behavior, and remove the old app/tests.rs file.

Closes #405
